pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pairbot"
version = "0.0.1"
readme = "README.md"
dependencies = [
  "SQLAlchemy",
  "alembic",
  "discord",
  "python-dotenv",
  "dateparser"
]

[project.urls]
Documentation = "https://github.com/scippage/pairbot#readme"
Issues = "https://github.com/scippage/pairbot/issues"
Source = "https://github.com/scippage/pairbot"

[project.scripts]
pairbot = "pairbot:run"

# Testing
[tool.hatch.envs.default]
python = "3.11"
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[tool.coverage.run]
source_pkgs = ["pairbot", "tests"]
branch = true
parallel = true
omit = []

[tool.coverage.paths]
pairbot = ["pairbot", "*/pairbot/pairbot"]
tests = ["tests", "*/pairbot/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

# Linting and formatting
[tool.hatch.envs.style]
detached = true
dependencies = [
  "black",
  "isort"
]

[tool.hatch.envs.style.scripts]
check = [
  "black --check --diff .",
  "isort --check-only --diff .",
]
format = [
  "isort .",
  "black .",
]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
profile = "black"
